+++ spec/dummy/app/models/los_record.rb
@@ -1,4 +1,5 @@
 class LosRecord < ActiveRecord::Base
   synced strategy: :updated_since, timestamp_strategy: Synced::Strategies::SyncedPerScopeTimestampStrategy
+
   belongs_to :account
 end