+++ spec/spec_helper.rb
@@ -23,6 +23,9 @@ RSpec.configure do |config|
   config.use_transactional_fixtures = true
   config.infer_base_class_for_anonymous_controllers = false
   config.order = "random"
+
+  config.filter_run focus: true
+  config.run_all_when_everything_filtered = true
 end
 
 def remote_object(attributes = {})