+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 * [feature] Allow to have a transaction per page of fetched objects by using :transaction_per_page option
+* [enhancement] Support only Ruby 2.3+
 
 ## 1.5.2
   * [bugfix] Fix n + 1 querry problems