+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## master
+## 1.6.0
 
 * [feature] Allow to have a transaction per page of fetched objects by using :transaction_per_page option
 * [enhancement] Support only Ruby 2.3+